Split main window code-behind partials
