+++ SentryReplay/MainWindow.About.cs
@@ -1,0 +1,12 @@
+﻿using CommunityToolkit.Mvvm.Input;
+
+namespace SentryReplay;
+
+public partial class MainWindow
+{
+    [RelayCommand]
+    private void ToggleAbout()
+    {
+        ShowAboutPage = !ShowAboutPage;
+    }
+}
